Implement CRUD for schedule entries
The whole structure of the schedule and schedule data changed to support
schedule entries. All the aggregations and getting schedule data now
uses the new schedule entry structure.

We don't use person and hourly wage from the client anymore as all
these settings are stored with the schedule. And when generating a
schedule, the person is coming from the UI. So ideally global settings
are obsolete now.